Replce VerticalBarGraphs with ParticipantAgeDistribution component
