+++ server/src/routes/index.ts
@@ -6,7 +6,6 @@ import GunRouter from './Gun';
 import ParticipantRouter from './Participant';
 import BarGraphsRouter from './BarGraphs';
 import LineGraphsRouter from './LineGraphs';
-import VerticalBarGraphsRouter from './VerticalBarGraphs';
 import StateComparisonsRouter from './StateComparisons';
 
 // Init router and path
@@ -21,7 +20,6 @@ router.use('/participant', ParticipantRouter);
 
 router.use('/linegraphs', LineGraphsRouter);
 router.use('/bargraphs', BarGraphsRouter);
-router.use('/verticalbargraphs', VerticalBarGraphsRouter);
 router.use('/statecomparisons', StateComparisonsRouter);
 
 // Export the base-router
